# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Refúgio da Ferradura API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Blog posts and articles
  - name: places
    description: Places and locations
  - name: media
    description: Media upload
  - name: auth
    description: Authentication
  - name: ai
    description: AI content generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all published posts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"

  /posts/admin:
    get:
      operationId: listPostsAdmin
      tags: [posts]
      summary: List all posts for admin (includes drafts)
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"

  /posts/{slug}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /posts/admin/{id}:
    get:
      operationId: getPostAdmin
      tags: [posts]
      summary: Get post by ID for admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /posts/admin/create:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /places:
    get:
      operationId: listPlaces
      tags: [places]
      summary: List all places
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of places
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceListResponse"

  /places/admin:
    get:
      operationId: listPlacesAdmin
      tags: [places]
      summary: List all places for admin
      responses:
        "200":
          description: List of places
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceListResponse"

  /places/{slug}:
    get:
      operationId: getPlace
      tags: [places]
      summary: Get place by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Place detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /places/admin/{id}:
    get:
      operationId: getPlaceAdmin
      tags: [places]
      summary: Get place by ID for admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlace
      tags: [places]
      summary: Update a place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaceBody"
      responses:
        "200":
          description: Updated place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePlace
      tags: [places]
      summary: Delete a place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /places/admin/create:
    post:
      operationId: createPlace
      tags: [places]
      summary: Create a new place
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaceBody"
      responses:
        "201":
          description: Created place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"

  /media/upload:
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload media file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Uploaded file info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"

  /ai/generate-from-url:
    post:
      operationId: generateFromUrl
      tags: [ai]
      summary: Generate article from URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateFromUrlBody"
      responses:
        "200":
          description: Generated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedArticle"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - success
        - user

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required:
        - id
        - username

    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - content
        - category
        - status
        - createdAt
        - updatedAt

    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
      required:
        - posts
        - total

    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        metaDescription:
          type: ["string", "null"]
      required:
        - title
        - slug
        - content
        - category
        - status

    UpdatePostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        metaDescription:
          type: ["string", "null"]

    Place:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        coverImage:
          type: ["string", "null"]
        images:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        featured:
          type: boolean
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - description
        - category
        - featured
        - createdAt
        - updatedAt

    PlaceListResponse:
      type: object
      properties:
        places:
          type: array
          items:
            $ref: "#/components/schemas/Place"
      required:
        - places

    CreatePlaceBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        coverImage:
          type: ["string", "null"]
        images:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        featured:
          type: boolean
        metaDescription:
          type: ["string", "null"]
      required:
        - name
        - slug
        - description
        - category

    UpdatePlaceBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        coverImage:
          type: ["string", "null"]
        images:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        featured:
          type: boolean
        metaDescription:
          type: ["string", "null"]

    MediaUploadResponse:
      type: object
      properties:
        url:
          type: string
        filename:
          type: string
      required:
        - url
        - filename

    GenerateFromUrlBody:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    GeneratedArticle:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
      required:
        - title
        - excerpt
        - content
        - category