# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Refúgio da Ferradura API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  username: string;
}