# lib/api-zod/src/generated/types/mediaUploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Refúgio da Ferradura API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaUploadResponse {
  url: string;
  filename: string;
}