# lib/api-zod/src/generated/types/place.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Refúgio da Ferradura API
 * OpenAPI spec version: 0.1.0
 */

export interface Place {
  id: number;
  name: string;
  slug: string;
  description: string;
  category: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  images?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  openingHours?: string | null;
  featured: boolean;
  /** @nullable */
  metaDescription?: string | null;
  createdAt: Date;
  updatedAt: Date;
}