# lib/api-zod/src/generated/types/updatePostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Refúgio da Ferradura API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePostBody {
  title?: string;
  slug?: string;
  /** @nullable */
  excerpt?: string | null;
  content?: string;
  /** @nullable */
  coverImage?: string | null;
  category?: string;
  status?: string;
  /** @nullable */
  metaDescription?: string | null;
}